requirements.txt
Django==3.2.20
drf-yasg==1.21.3
djangorestframework==3.13.1
django-model-utils==4.2.0
django-extensions==3.2.0
psycopg2-binary==2.9.3
gunicorn==20.1.0
coverage[toml]==6.5.0
requests==2.28.1
requests-mock==1.9.3
django-filter==22.1
django-qr-code==2.3.0
django-ses==3.1.2
boto3==1.24.57
Pillow==9.3.0
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-mock==3.8.2
freezegun==1.2.2
marshmallow==3.18.0
django-post_office==3.6.1
factory-boy==3.2.1
Faker==14.1.0
django-structlog==3.0.1
qrcode==7.3.1
